+++ tester/src/lib.rs
@@ -5,4 +5,4 @@ pub mod checks;
 pub mod context;
 pub mod error;
 pub mod fixtures;
-pub mod transactions;
+pub mod steps;
